moves Vector and Exponent tests cases to test_algebra.py module
