+++ tests/ucon/test_dimension.py
@@ -1,33 +1,5 @@
 import unittest
-from ucon.dimension import Vector, Dimension
-
-
-class TestVector(unittest.TestCase):
-
-    def test_vector_iteration_and_length(self):
-        v = Vector(1, 0, 0, 0, 0, 0, 0)
-        self.assertEqual(tuple(v), (1, 0, 0, 0, 0, 0, 0))
-        self.assertEqual(len(v), 7)  # always 7 components
-
-    def test_vector_addition(self):
-        v1 = Vector(1, 0, 0, 0, 0, 0, 0)
-        v2 = Vector(0, 2, 0, 0, 0, 0, 0)
-        result = v1 + v2
-        self.assertEqual(result, Vector(1, 2, 0, 0, 0, 0, 0))
-
-    def test_vector_subtraction(self):
-        v1 = Vector(2, 1, 0, 0, 0, 0, 0)
-        v2 = Vector(1, 1, 0, 0, 0, 0, 0)
-        self.assertEqual(v1 - v2, Vector(1, 0, 0, 0, 0, 0, 0))
-
-    def test_vector_equality_and_hash(self):
-        v1 = Vector(1, 0, 0, 0, 0, 0, 0)
-        v2 = Vector(1, 0, 0, 0, 0, 0, 0)
-        v3 = Vector(0, 1, 0, 0, 0, 0, 0)
-        self.assertTrue(v1 == v2)
-        self.assertFalse(v1 == v3)
-        self.assertEqual(hash(v1), hash(v2))
-        self.assertNotEqual(hash(v1), hash(v3))
+from ucon.dimension import Dimension
 
 
 class TestDimension(unittest.TestCase):
@@ -108,58 +80,6 @@ class TestDimension(unittest.TestCase):
         self.assertNotEqual(Dimension.mass.value, Dimension.time.value)
         self.assertEqual(Dimension.mass, Dimension.mass)
         self.assertNotEqual(Dimension.mass, Dimension.time)
-
-
-class TestVectorEdgeCases(unittest.TestCase):
-
-    def test_zero_vector_equality_and_additivity(self):
-        zero = Vector()
-        self.assertEqual(zero, Vector(0, 0, 0, 0, 0, 0, 0))
-        # Adding or subtracting zero should yield same vector
-        v = Vector(1, 2, 3, 4, 5, 6, 7)
-        self.assertEqual(v + zero, v)
-        self.assertEqual(v - zero, v)
-
-    def test_vector_with_negative_exponents(self):
-        v1 = Vector(1, -2, 3, 0, 0, 0, 0)
-        v2 = Vector(-1, 2, -3, 0, 0, 0, 0)
-        result = v1 + v2
-        self.assertEqual(result, Vector(0, 0, 0, 0, 0, 0, 0))
-        self.assertEqual(v1 - v1, Vector())  # perfect cancellation
-
-    def test_vector_equality_with_non_vector(self):
-        v = Vector()
-        with self.assertRaises(AssertionError):
-            v == "not a vector"
-        with self.assertRaises(AssertionError):
-            v == None
-
-    def test_hash_consistency_for_equal_vectors(self):
-        v1 = Vector(1, 0, 0, 0, 0, 0, 0)
-        v2 = Vector(1, 0, 0, 0, 0, 0, 0)
-        self.assertEqual(hash(v1), hash(v2))
-        self.assertEqual(len({v1, v2}), 1)
-
-    def test_iter_length_order_consistency(self):
-        v = Vector(1, 2, 3, 4, 5, 6, 7)
-        components = list(v)
-        self.assertEqual(len(components), len(v))
-        # Ensure order of iteration is fixed (T→L→M→I→Θ→J→N)
-        self.assertEqual(components, [1, 2, 3, 4, 5, 6, 7])
-
-    def test_vector_arithmetic_does_not_mutate_operands(self):
-        v1 = Vector(1, 0, 0, 0, 0, 0, 0)
-        v2 = Vector(0, 1, 0, 0, 0, 0, 0)
-        _ = v1 + v2
-        self.assertEqual(v1, Vector(1, 0, 0, 0, 0, 0, 0))
-        self.assertEqual(v2, Vector(0, 1, 0, 0, 0, 0, 0))
-
-    def test_invalid_addition_type_raises(self):
-        v = Vector(1, 0, 0, 0, 0, 0, 0)
-        with self.assertRaises(TypeError):
-            _ = v + "length"
-        with self.assertRaises(TypeError):
-            _ = v - 5
 
 
 class TestDimensionEdgeCases(unittest.TestCase):
